+++ integrations/azure/azure_integration/utils.py
@@ -2,13 +2,18 @@ import contextlib
 import enum
 import typing
 
+import aiostream
+
 from port_ocean.context.event import event
-from port_ocean.context.ocean import ocean
 from azure.identity.aio import DefaultAzureCredential
 from azure.mgmt.resource.resources.v2022_09_01.aio import ResourceManagementClient
 
-from azure_integration.overrides import AzureResourceConfig
-
+from azure_integration.overrides import (
+    AzureSpecificKindsResourceConfig,
+    AzureCloudResourceConfig,
+    AzureCloudResourceSelector,
+    AzureSpecificKindSelector,
+)
 
 BATCH_SIZE = 20
 
@@ -20,15 +25,13 @@ class ResourceKindsWithSpecialHandling(enum.StrEnum):
     """
 
     RESOURCE_GROUPS = "Microsoft.Resources/resourceGroups"
+    SUBSCRIPTION = "subscription"
+    CLOUD_RESOURCE = "cloudResource"
 
 
-def get_integration_subscription_id() -> str:
-    logic_settings = ocean.integration_config
-    subscription_id = logic_settings["subscription_id"]
-    return subscription_id
-
-
-def get_current_resource_config() -> AzureResourceConfig:
+def get_current_resource_config() -> (
+    typing.Union[AzureSpecificKindsResourceConfig, AzureCloudResourceConfig]
+):
     """
     Returns the current resource config, accessible only inside an event context
     """
@@ -71,6 +74,36 @@ def resolve_resource_type_from_resource_uri(resource_uri: str) -> str:
                 resource_type += "/" + resource[9:][resource_kind_extension]
 
     return resource_type
+
+
+def get_resource_configs_with_resource_kind(
+    resource_kind: str,
+    resource_configs: typing.List[
+        typing.Union[AzureSpecificKindsResourceConfig, AzureCloudResourceConfig]
+    ],
+) -> typing.List[
+    typing.Union[AzureSpecificKindsResourceConfig, AzureCloudResourceConfig]
+]:
+    """
+    Returns the resource configs that have the resource kind
+
+    :param resource_kind: Resource kind
+    :param resource_configs: List of resource configs
+    :return: List of resource configs that have the resource kind
+    """
+    return [
+        resource_config
+        for resource_config in resource_configs
+        if (
+            resource_config.kind == resource_kind
+            and isinstance(resource_config.selector, AzureSpecificKindSelector)
+        )
+        or (
+            resource_config.kind == ResourceKindsWithSpecialHandling.CLOUD_RESOURCE
+            and isinstance(resource_config.selector, AzureCloudResourceSelector)
+            and resource_kind in resource_config.selector.resource_kinds.keys()
+        )
+    ]
 
 
 def is_sub_resource(resource_type: str) -> bool:
@@ -129,13 +162,30 @@ async def batch_resources_iterator(
 
 
 @contextlib.asynccontextmanager
-async def resource_client_context() -> typing.AsyncIterator[ResourceManagementClient]:
+async def resource_client_context(
+    subscription_id: str,
+) -> typing.AsyncIterator[ResourceManagementClient]:
     """
     Creates a resource client context manager that yields a resource client with the default azure credentials
     """
     async with DefaultAzureCredential() as credential:
         async with ResourceManagementClient(
             credential=credential,
-            subscription_id=get_integration_subscription_id(),
+            subscription_id=subscription_id,
         ) as client:
             yield client
+
+
+async def stream_async_iterators_tasks(
+    tasks: typing.List[typing.AsyncIterable[typing.Any]],
+) -> typing.AsyncIterable[typing.Any]:
+    """
+    Streams the results of multiple async iterators
+
+    :param tasks: A list of async iterators
+    :return: A stream of results
+    """
+    combine = aiostream.stream.merge(tasks[0], *tasks[1:])
+    async with combine.stream() as streamer:
+        async for batch_items in streamer:
+            yield batch_items